fix(port): compile 5 all-features test errors exposed by phase-1 gates job
The Phase-1 carrier gates job runs cargo test --all-features, the first time
the port lineage has ever executed it. Three port seams were compile-broken
under that set (survivor unsubscribe-drop-free path, single telegram runs
never compiled the other features):

- bg_resume.rs:166  assert_eq!(usize, u64)  -> cast both to usize
- telegram_mermaid_test.rs:351  build_body_markdown_media now takes reply_to
  (5th arg) since the #1238 ladder -> pass None
- telegram_rich_api_test.rs  MediaEntry gained fields bytes + url: Option
  since the bytes attach:// path -> url: Some(..), bytes: None

No behavioral change to shipped binary (telegram-only compiled fine already).

Session-Id: e756b84b-b645-43d2-bfb6-a903f95a41f4
[role-exception] port fixup (all-features gate)

+++ src/channels/bg_resume.rs
@@ -163,7 +163,7 @@ mod tests {
         )
         .await;
         assert_eq!(got, None);
-        assert_eq!(polls.load(Ordering::Relaxed), READY_WAIT_SECS);
+        assert_eq!(polls.load(Ordering::Relaxed) as usize, READY_WAIT_SECS as usize);
     }
 
     /// Already-ready handle → immediate delivery, single poll, no sleep.
